refactor: Modularize executor.rs into focused module hierarchy (Issue #1145)
Split the 824-line executor.rs into a cleaner module structure:

## Module Structure
- mod.rs (163 lines): Core SqlExecutor struct and execute() logic
- loader.rs (88 lines): Database loading from SQL dumps
- validation.rs (125 lines): Input validation for security
- copy_handler.rs (78 lines): COPY command import/export handling
- display.rs (276 lines): Table description and list operations
- tests.rs (132 lines): All test cases

## Benefits
- **Separation of concerns**: Each module has single responsibility
- **Testability**: Components can be tested independently
- **Maintainability**: Files under 300 lines, easier to navigate
- **Security**: Validation logic isolated and easier to audit
- **Reusability**: Display formatting available for other CLI tools

## Testing
All 50 tests pass successfully:
- Executor core functionality
- Validation (SQL injection prevention)
- Display operations (describe, list commands)
- Row count reporting

No API changes - SqlExecutor methods remain publicly accessible.

Closes #1145
