+++ crates/cli/src/executor/tests.rs
@@ -1,0 +1,130 @@
+use super::{validation, SqlExecutor};
+
+#[test]
+fn test_list_schemas() {
+    let executor = SqlExecutor::new(None).unwrap();
+    // Default database should have "public" schema
+    assert!(executor.list_schemas().is_ok());
+}
+
+#[test]
+fn test_list_indexes_empty() {
+    let executor = SqlExecutor::new(None).unwrap();
+    // New database should have no indexes
+    assert!(executor.list_indexes().is_ok());
+}
+
+#[test]
+fn test_list_roles() {
+    let executor = SqlExecutor::new(None).unwrap();
+    // Should show at least the default PUBLIC role
+    assert!(executor.list_roles().is_ok());
+}
+
+#[test]
+fn test_validate_table_name_nonexistent() {
+    let executor = SqlExecutor::new(None).unwrap();
+    // Should fail for non-existent table
+    let result = validation::validate_table_name(&executor.db, "nonexistent_table");
+    assert!(result.is_err());
+    assert!(result.unwrap_err().to_string().contains("does not exist"));
+}
+
+#[test]
+fn test_validate_table_name_sql_injection() {
+    let executor = SqlExecutor::new(None).unwrap();
+    // Should fail for table names with SQL injection attempts
+    let result = validation::validate_table_name(&executor.db, "users; DROP TABLE users; --");
+    assert!(result.is_err());
+}
+
+#[test]
+fn test_describe_table_basic() {
+    let mut executor = SqlExecutor::new(None).unwrap();
+    executor.execute("CREATE TABLE test (id INT PRIMARY KEY, name VARCHAR(50))").unwrap();
+    // Should print table description without error
+    assert!(executor.describe_table("test").is_ok());
+}
+
+#[test]
+fn test_describe_nonexistent_table() {
+    let executor = SqlExecutor::new(None).unwrap();
+    let result = executor.describe_table("nonexistent");
+    assert!(result.is_err());
+    assert!(result.unwrap_err().to_string().contains("does not exist"));
+}
+
+#[test]
+fn test_describe_table_with_indexes() {
+    let mut executor = SqlExecutor::new(None).unwrap();
+    executor.execute("CREATE TABLE test (id INT PRIMARY KEY, email VARCHAR(100))").unwrap();
+    // Note: CREATE INDEX is not yet supported in CLI executor, so skip for now
+    // This test will pass once CREATE INDEX support is added
+    assert!(executor.describe_table("test").is_ok());
+}
+
+#[test]
+fn test_describe_table_with_multiple_columns() {
+    let mut executor = SqlExecutor::new(None).unwrap();
+    executor.execute("CREATE TABLE products (id INT PRIMARY KEY, name VARCHAR(100), price DECIMAL(10, 2))").unwrap();
+    // Should print table with multiple columns of different types
+    assert!(executor.describe_table("products").is_ok());
+}
+
+#[test]
+fn test_insert_row_count_single() {
+    let mut executor = SqlExecutor::new(None).unwrap();
+    executor.execute("CREATE TABLE users (id INT PRIMARY KEY, name VARCHAR(50))").unwrap();
+
+    let result = executor.execute("INSERT INTO users (id, name) VALUES (1, 'Alice')").unwrap();
+    assert_eq!(result.row_count, 1, "Single INSERT should return row count of 1");
+}
+
+#[test]
+fn test_insert_row_count_multiple() {
+    let mut executor = SqlExecutor::new(None).unwrap();
+    executor.execute("CREATE TABLE users (id INT PRIMARY KEY, name VARCHAR(50))").unwrap();
+
+    let result = executor.execute(
+        "INSERT INTO users (id, name) VALUES (1, 'Alice'), (2, 'Bob'), (3, 'Charlie')"
+    ).unwrap();
+    assert_eq!(result.row_count, 3, "Multiple value INSERT should return row count of 3");
+}
+
+#[test]
+fn test_update_row_count() {
+    let mut executor = SqlExecutor::new(None).unwrap();
+    executor.execute("CREATE TABLE users (id INT PRIMARY KEY, name VARCHAR(50))").unwrap();
+    executor.execute("INSERT INTO users (id, name) VALUES (1, 'Alice'), (2, 'Bob'), (3, 'Charlie')").unwrap();
+
+    let result = executor.execute("UPDATE users SET name = 'Updated' WHERE id > 1").unwrap();
+    assert_eq!(result.row_count, 2, "UPDATE should return row count of 2");
+}
+
+#[test]
+fn test_delete_row_count() {
+    let mut executor = SqlExecutor::new(None).unwrap();
+    executor.execute("CREATE TABLE users (id INT PRIMARY KEY, name VARCHAR(50))").unwrap();
+    executor.execute("INSERT INTO users (id, name) VALUES (1, 'Alice'), (2, 'Bob'), (3, 'Charlie')").unwrap();
+
+    let result = executor.execute("DELETE FROM users WHERE id IN (1, 3)").unwrap();
+    assert_eq!(result.row_count, 2, "DELETE should return row count of 2");
+}
+
+#[test]
+fn test_select_row_count() {
+    let mut executor = SqlExecutor::new(None).unwrap();
+    executor.execute("CREATE TABLE users (id INT PRIMARY KEY, name VARCHAR(50))").unwrap();
+    executor.execute("INSERT INTO users (id, name) VALUES (1, 'Alice'), (2, 'Bob')").unwrap();
+
+    let result = executor.execute("SELECT * FROM users").unwrap();
+    assert_eq!(result.row_count, 2, "SELECT should return row count of 2");
+    assert_eq!(result.rows.len(), 2, "SELECT should return 2 rows");
+}
+
+#[test]
+fn test_create_table_row_count() {
+    let mut executor = SqlExecutor::new(None).unwrap();
+    let result = executor.execute("CREATE TABLE test (id INT PRIMARY KEY)").unwrap();
+    assert_eq!(result.row_count, 0, "CREATE TABLE should return row count of 0 (DDL)");
+}
